rename detailsScreen package name
